requirements.txt
streamlit
google.generativeai
pyPDF2
python-dotenv